Create TrappingRainWater
Here we have to find trapped rainwater. So here first we gonna find the maximum bar from the left side and also from the right side. We have to create a boundary. Now we have to take the minimum between the left side and the right side. Because we need a minimum one otherwise we can not store water. After this, we have to subtract the height with a minimum one and we will get our result.
TrappedWater = MIN(MAXofLEFT,MAXofRIGHT) - HEIGHT(ARRAYS ELEMENT)