Create Find the Student that Will Replace the Chalk
Lets take an example arr[] = {5,1,5}. and k is 22
Now , 22 - 5 = 17
          17 - 1 = 16
          16 - 5 = 11
then again, 
          11 - 5 = 6
          6 - 1 = 5
          5 - 5 = 0
Now my k is zero current element is 0 which requires 5 chalks so not possible so ans is 0

So here if I calculates the total sum that is 11. Now the exp is 11*2 = 22 , so there is no remaining and k = 22 and sum = 11 modulus is 0. So 1st element requirement is 5 but I cant give as modulus is 0(remaining) . if suppose , remaining is 5 then I can give the required chalks to the 1st person and then ans would be index 1