Create MajorityElementII
Algorithm: First i have taken ele1,ele2,count1,count2.
Now take this example
3 2 3
So first, ele1 will be 3,count1 = 1
             ele2 will be 2,count2 = 1
then,     for last element 3 as i have already count1 !=0 and count2!=0 so 3rd else if will executed count1 will be 2. Now i will be checking which one is greater than N/3. And one thing for first two if condition i have to check that ele1!=nums[i] for second if and ele2!=nums[i] for First if so that my ele1 and ele2 are not same.