Create Longest Substring Without Repeating Characters
So here i have taken one hashmap where the key is the character and the value is the index of the character
Suppose String is = "abcabcbb"
Here in Hashmap first I put a->0,b->1,c->2, then I got 'a' that is already present in the map. So now I will start from the b->1, but before starting from the b I will store map.size() into max for future maximum comparison.