Create Next Greater Element II
hypothetically, I just doubled the array . Now
for example. 1 2 3 4 3,
1 2 3 4 3 1 2 3 4 3
Now lets start from last. i will add 3 in the stack. next is 4 which is greater than 3 so remove 3. Add 4 in stack(4) is present. Now again 3 no problem (3 4), next 2 and then 1 so stack(1 2 3 4) now again i got 3 and here my index is less than ArrayLength. From here i will find the next greater element. So 3 is less than 1 2 3 , so remove this and add 3 so now stack is (3,4) so here i got my 3's NG is 4, Now i got 4 which is less than and equal to present element in stack. so remove it and  add 4 to the stack(4). Before entering as stack was empty so we gonna add (-1)
then add 2 no problem stack is (2,3,4) so here i got my 2'S NG is 3,like that for one (1,2,3,4) so here i got 1'S NG is 2.