Create Find the Prefix Common Array of Two Arrays
i have taken two hashsets and check that if i have the B's elements in set1 or not. if i got increment the counter. same way i did for A's elements in set2. Now if i got same element then i decrement the counter by 1. This is how i solved the problem.